+++ homeassistant/components/tradfri/fan.py
@@ -25,7 +25,16 @@ from .const import (
     DOMAIN,
     KEY_API,
 )
-from .utils import _from_fan_percentage, _from_fan_speed
+
+
+def _from_fan_percentage(percentage: int) -> int:
+    """Convert percent to a value that the Tradfri API understands."""
+    return round(percentage / 100 * ATTR_MAX_FAN_STEPS)
+
+
+def _from_fan_speed(fan_speed: int) -> int:
+    """Convert the Tradfri API fan speed to a percentage value."""
+    return round(fan_speed / ATTR_MAX_FAN_STEPS * 100)
 
 
 async def async_setup_entry(
